% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padma_data.R
\docType{data}
\name{mirtarbase}
\alias{mirtarbase}
\title{Curated miR-target interaction predictions from miRTarBase}
\format{
An object of class \code{data.frame} with two
columns: \code{miRNA}, which provides the miRNA identifier
(e.g., \code{'hsa-miR-20a-5p'}) and
\code{Target Gene}, which provides the corresponding
predicted gene target.
}
\source{
\url{http://mirtarbase.mbc.nctu.edu.tw/php/index.php}
}
\usage{
data(mirtarbase)
}
\description{
Data.frame of 10,754 predicted miRNA gene targets from miRTarBase
(version 7.0), filtered to include only predictions with the
'Functional MTI' support type.
}
\examples{
data(mirtarbase)
head(mirtarbase)
}
\references{
Chou et al. (2018) Nucleic Acids Research 46, D296-D302.
\url{https://doi.org/10.1093/nar/gkx1067}.
}
\keyword{datasets}
