% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{padma}
\alias{padma}
\title{Calculate individualized deviation scores from multi-omic data}
\usage{
padma(omics_data, base_ids = NULL, supp_ids = NULL, apply_log = NULL,
  pathway_name = "c2_cp_BIOCARTA_D4GDI_PATHWAY",
  impute_MFA_missMDA = FALSE, mirna_targets = NULL,
  full_results = TRUE)
}
\arguments{
\item{omics_data}{Named list (including "clinical" and "rnaseq", and optionally
additionaly including one or more of "cna", "methyl", and/or "mirna"),
where each element of the list correspnds to a data.frame with
omics data from n matched individuals. The data.frame for clinical
data ("clinical") has dimension n x (q + 1) for q clinical variables, where an additional
column named "bcr_patient_barcode" contains unique TCGA sample identifiers.
The data.frames for
RNA-seq ("rnaseq"), copy number alterations ("cna"), and methylation ("methyl") have
dimension n x (p + 1) for p genes, where the first column contains gene names. The
data.frame for microRNA-seq ("mirna") has dimension n x (p + 2) for p genes, where the
first column contains the miRNA ID and the second contains the gene name of predicted
targets. Row names of all data.frames should contain the same unique sample identifiers
as those included in the "bcr_patient_barcode" variable of the "clinical" data.frame.
Sample orders (rows of "clinical" and columns of the others) should match across all
data.frames}

\item{base_ids}{(optional) Sample indices to be used as reference base data.
By default, all samples are used.}

\item{supp_ids}{(optional) Sample indices to be used as supplementary individuals to be
projected onto the analysis based on the individuals identified in \code{base_ids}. By
default, takes the value \code{NULL}, but should not overlap with \code{base_ids} if
provided by the user.}

\item{apply_log}{(optional) Data types (corresponding to the names of the \code{omics_data}
list) to which a log + 1 transformation should be applied.}

\item{pathway_name}{Either a character of a KEGG pathway identifier or MSigDB pathway names
(e.g., see the pathway names in the \code{"geneset"} column of the preloaded
\code{msigdb} data in the package), or a vector of gene symbols.}

\item{impute_MFA_missMDA}{If \code{TRUE}, impute missing values separately in base and
supplementary data using MFA as implemented in the \emph{missMDA} package; otherwise mean
imputation is used (default).}

\item{mirna_targets}{(optional) Data.frame with 2 columns: column \code{"miRNA"}
should provide a miRNA name (e.g., \code{"hsa-miR-20a-5p"}) and column \code{"Target Gene"}
should provide the symbol of a predicted gene target for the corresponding miRNA.
Each predicted miRNA-target pair should have its own row in the
data.frame. By default, the miRNA-gene interactions of type "Functional MTI" from miRTarBase
are used (see the preloaded \code{"mirtarbase} data in the package).}

\item{full_results}{If \code{TRUE} (default), include full MFA results in function output;
otherwise, provide concise output to save space.}
}
\value{
Output of class \code{"padma"} containing a list of several elements.
}
\description{
Calculate individualized deviation scores from multi-omic data
}
\examples{

run_padma <- padma(LUAD_subset,
                   pathway_name = "c2_cp_BIOCARTA_D4GDI_PATHWAY")

plot_factor_map(run_padma, dim_x = 1, dim_y = 2)
plot_partial_factor_map(run_padma, dim_x = 1, dim_y = 2)
plot_partial_factor_map(run_padma, id = "TCGA-78-7536", dim_x = 1, dim_y = 2)
plot_omics_contrib(run_padma, max_dim = 10)
}
