% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{padmaRun}
\alias{padmaRun}
\title{Calculate individualized deviation scores from multi-omic data}
\usage{
padmaRun(
  omics_data,
  gene_map = padma::mirtarbase,
  base_ids = NULL,
  supp_ids = NULL,
  pathway_name = "c2_cp_BIOCARTA_D4GDI_PATHWAY",
  impute = FALSE,
  variance_threshold = 1e-04,
  full_results = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{omics_data}{Object of class \code{'MultiAssayExperiment'} containing 
omics data from n matched individuals.}

\item{gene_map}{(optional) Data frame mapping
arbitrary biological entities (e.g. miRNAs) to genes. Contains two columns, 
where the first provides the IDs of the entity and
the second provides the IDs of the corresponding target gene.
By default, the miRNA-gene interactions of type 'Functional MTI' from 
miRTarBase are used (see the preloaded \code{'mirtarbase'} data in the 
package).}

\item{base_ids}{(optional) Sample names to be used as reference base data.
By default, all samples are used.}

\item{supp_ids}{(optional) Sample names to be used as supplementary 
individuals to be
projected onto the analysis based on the individuals identified in 
\code{base_ids}. By default, takes the value \code{NULL}, but should not 
overlap with \code{base_ids} if provided by the user.}

\item{pathway_name}{Character of either a KEGG pathway identifier or MSigDB 
pathway names (e.g., see the pathway names in the \code{'geneset'} column of 
the preloaded \code{msigdb} data in the package), or a vector of gene 
symbols.}

\item{impute}{If \code{TRUE}, impute missing values separately in base and
supplementary data using MFA as implemented in the \emph{missMDA} package; 
otherwise simple mean imputation is used (default).}

\item{variance_threshold}{Minimal variance required across samples to retain 
a biological entity in the analysis}

\item{full_results}{If \code{TRUE} (default), include full MFA results in 
function output; otherwise, provide concise output to save space.}

\item{verbose}{If \code{TRUE}, provide verbose output.}

\item{...}{Optional additional arguments}
}
\value{
An S4 object of class \code{padmaResults}, where individualized pathway 
deviation scores are stored as the assay data, and the corresponding 
{pathway name, full MFA results, number of genes, and names of imputed 
or filtered genes} are stored as slots that can be retrieved using
the appropriate accessor functions.
}
\description{
Calculate individualized deviation scores from multi-omic data
}
\examples{

LUAD_subset <- padma::LUAD_subset
## Create MultiAssayExperiment object with LUAD data
omics_data <- 
  list(rnaseq = LUAD_subset$rnaseq,
       methyl = LUAD_subset$methyl,
       mirna = LUAD_subset$mirna,
       cna = LUAD_subset$cna)
pheno_data <- 
  data.frame(LUAD_subset$clinical, 
             row.names = LUAD_subset$clinical$bcr_patient_barcode)
mae <-
  suppressMessages(
    MultiAssayExperiment::MultiAssayExperiment(
      experiments = omics_data, colData = pheno_data))

## Run padma
run_padma <- 
  padma(mae, gene_map = padma::mirtarbase,
        pathway_name = "c2_cp_BIOCARTA_D4GDI_PATHWAY", verbose = FALSE)

summary(run_padma)

## padma plots
factorMap(run_padma, dim_x = 1, dim_y = 2)
factorMap(run_padma, dim_x = 1, dim_y = 2,
           partial_id = "TCGA-78-7536")
omicsContrib(run_padma, max_dim = 10)
}
