% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_factor_map}
\alias{plot_factor_map}
\title{Plot an MFA factor map for individuals based on padma analysis}
\usage{
plot_factor_map(padma_obj, dim_x = 1, dim_y = 2, plot_ellipse = TRUE)
}
\arguments{
\item{padma_obj}{Output from running the \code{padma} function
(with \code{"full_results = TRUE"})}

\item{dim_x}{Dimension number of the MFA to be plotted on the x-axis.}

\item{dim_y}{Dimension number of the MFA to be plotted on the y-axis.}

\item{plot_ellipse}{If \code{TRUE}, superimpose a normal confidence ellipsis
on the factor map.}
}
\value{
Factor map of class \code{ggplot}.
}
\description{
Produce an MFA factor map for individuals for a pair of dimensions provided by the user, where
a normal confidence ellipsis is optionally superimposed.
}
