% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_omics_contrib}
\alias{plot_omics_contrib}
\title{Plot the omics contribution per MFA axis and the overall weighted contribution}
\usage{
plot_omics_contrib(padma_obj, max_dim = min(10, nrow(padma_obj$eig)))
}
\arguments{
\item{padma_obj}{Output from running the \code{padma} function
(with \code{"full_results = TRUE"})}

\item{max_dim}{Maxomum imension number of the MFA to be plotted.}
}
\value{
Barplots of percent variance contribution of class \code{ggplot}.
}
\description{
Plot barplots indicating the percent contribution of each omics to each MFA dimension,
as well as the overall weighted (by eigenvalue) percent contribution to the full
analysis.
}
