% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_partial_factor_map}
\alias{plot_partial_factor_map}
\title{Plot an MFA partial factor map for a given individual and genes based on padma analysis}
\usage{
plot_partial_factor_map(padma_obj,
  id = padma_obj$pathway_deregulation$bcr_patient_barcode[1],
  dim_x = 1, dim_y = 2)
}
\arguments{
\item{padma_obj}{Output from running the \code{padma} function
(with \code{"full_results = TRUE"})}

\item{id}{Index or sample name to be plotted. By default, the first sample is plotted.}

\item{dim_x}{Dimension number of the MFA to be plotted on the x-axis.}

\item{dim_y}{Dimension number of the MFA to be plotted on the y-axis.}
}
\value{
Partial factor map of class \code{ggplot}.
}
\description{
Produce an MFA partial factor map for a given individual for a pair of dimensions provided
by the user.
}
