% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padma_data.R
\docType{data}
\name{LUAD_subset}
\alias{LUAD_subset}
\title{Subset of batch-corrected multi-omic TCGA data in lung adenocarcinoma}
\format{A named list of five objects of class \code{data.frame}
containing a subset of the batch-corrected multi-omic TCGA
data from lung adenocarcinoma, corresponding to the 13 genes
in the D4 GDI signaling pathway:
\code{"clinical"} is of dimension 144 x 55 and contains clinical
variables for the 144 individuals. \code{"rnaseq"}, \code{"methyl"},
and \code{"cna"} are of dimension 13 (genes) x 145, where the
first column \code{"gene"} contains the gene symbol and the
remaining 144 columns correspond to the TCGA individuals.
\code{"mirna"} is of dimension 1 x 146 for the single miRNA
predicted to target one of the D4 GDI signaling pathway genes,
and the first two columns \code{miRNA} and \code{Target Gene}
respectively provide the miRNA identifier
(\code{"hsa-mir-421"}) and the corresponding
predicted gene target (\code{"hsa-mir-421"}).}
\source{
\href{https://www.cancer.gov/about-nci/organization/ccg/research/structural-genomics/tcga}{TCGA}
}
\usage{
data(LUAD_subset)
}
\description{
Multi-omic (RNA-seq, copy number alterations, methylation, and
miRNA-seq) and phenotypic data in 144 individuals in the TCGA-LUAD
data for the 13 genes in the D4-GDI signaling pathway.
}
\examples{
data(LUAD_subset)
head(LUAD_subset)
}
\references{
The Cancer Genome Atlas Research Network (2014) Nature 511, 543-550.
\href{https://doi.org/10.1038/nature13385}.

Rau et al. (2019) bioRxiv, \href{https://doi.org/10.1101/827022}.
}
\keyword{datasets}
